% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmiddens-package.R
\docType{package}
\name{webmiddens-package}
\alias{webmiddens-package}
\alias{webmiddens}
\title{Cache mocked HTTP requests}
\description{
Cache mocked HTTP requests
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
