% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wm_configuration.R
\name{wm_configuration}
\alias{wm_configuration}
\title{wm_configuration}
\usage{
wm_configuration(path, expire = NULL)
}
\arguments{
\item{path}{(character) path to set for the current midden}

\item{expire}{(numeric/integer) an expiration time (seconds)}
}
\value{
nothing, sets path and expire settings
}
\description{
webmiddens settings
}
\examples{
\dontrun{
wm_configuration("foobar")
wm_configuration("foobar", expire = 6)
}
}
