% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midden.R
\docType{data}
\name{midden}
\alias{midden}
\title{web middens}
\description{
web middens
}
\details{
\strong{Methods}
\describe{
\item{\code{init(path)}}{
a directory path
}
\item{\code{call(...)}}{
an http request code block
}
}
}
\examples{
\dontrun{
library(crul)

# without middens
con <- crul::HttpClient$new("https://httpbin.org")
con$get(query = list(stuff = "bananas"))

# with middens
x <- midden$new()
x
x$init(path = "rainforest")
x
x$call(con$get("get", query = list(stuff = "bananas")))

x <- midden$new(verbose = TRUE)
x$init(path = "rainforest")
x$call(con$get("get", query = list(stuff = "bananas")))
}
}
\keyword{datasets}
