% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midden.R
\name{midden}
\alias{midden}
\title{web middens}
\description{
web middens

web middens
}
\details{
\strong{Methods}
\describe{
\item{\code{init(path)}}{
a directory path
}
\item{\code{call(..., expire = NULL)}}{
an http request code block
- ...: a http request block
- expire: (integer) number of seconds until expiry. after this time,
we force a real HTTP reqeuest even if a matching stub exists.
times are recorded in UTC.
}
}
}
\examples{
\dontrun{
library(crul)

# without middens
con <- crul::HttpClient$new("https://httpbin.org")
con2 <- crul::HttpClient$new("https://google.com")
con$get("get", query = list(stuff = "bananas"))
con2$get(query = list(q = "stuff"))

# with middens
x <- midden$new()
x
x$init(path = "rainforest3")
x$cache
x
# first request is a real HTTP request
x$call(con$get("get", query = list(stuff = "bananas")))
# following requests use the cached response
x$call(con$get("get", query = list(stuff = "bananas")))

# verbose output
x <- midden$new(verbose = TRUE)
x$init(path = "rainforest")
x$call(con$get("get", query = list(stuff = "bananas")))

# set expiration time
x <- midden$new()
x$init(path = "grass")
x
# set expiry
x$call(con$get("get", query = list(grass = "tall")), expire = 3)
## before expiry, get mocked response
x$call(con$get("get", query = list(grass = "tall")), expire = 3)
Sys.sleep(5)
## after expiry, get real response
x$call(con$get("get", query = list(grass = "tall")), expire = 3)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{midden$new()}}
\item \href{#method-print}{\code{midden$print()}}
\item \href{#method-call}{\code{midden$call()}}
\item \href{#method-init}{\code{midden$init()}}
\item \href{#method-destroy}{\code{midden$destroy()}}
\item \href{#method-expire}{\code{midden$expire()}}
\item \href{#method-clone}{\code{midden$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midden$new(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midden$print(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call"></a>}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midden$call(..., expire = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\subsection{Method \code{init()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midden$init(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midden$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-expire"></a>}}
\subsection{Method \code{expire()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midden$expire(time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{midden$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
