% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{midden_tools}
\alias{midden_tools}
\alias{midden_current}
\alias{midden_kill}
\alias{midden_path}
\title{midden_tools}
\usage{
midden_current()

midden_kill()

midden_path()
}
\description{
get the current midden
}
\section{Methods}{

\itemize{
\item \code{midden_current()}: get the current midden
\item \code{midden_path()}: gethe current midden path
\item \code{midden_kill()}: deletes the current midden
}
}

